refactor(Section): move XDSSection to its own module, extract SizeValue to utils
Move XDSSection from Layout/Container/ to Section/ for discoverability.
Extract SizeValue type to utils/types.ts as a shared type — it's used by
Card, Section, Center, and Grid, so it shouldn't live in any one component.

Changes:
- Create packages/core/src/Section/ with component, index, and README
- Create packages/core/src/utils/types.ts with SizeValue
- Add ./Section export to package.json
- Add Section to top-level src/index.ts barrel export
- Update Card, Center, Grid to import SizeValue from utils/types
- Remove XDSSection.tsx from Layout/Container/
- Update Container/index.ts to re-export from Section/ and Card/
- Update all storybook stories to import from '@xds/core/Section'
