+++ apps/storybook/stories/Card.stories.tsx
@@ -1,8 +1,8 @@
 import type {Meta, StoryObj} from '@storybook/react';
 import * as stylex from '@stylexjs/stylex';
 import {XDSCard} from '@xds/core/Card';
+import {XDSSection} from '@xds/core/Section';
 import {
-  XDSSection,
   XDSVStack,
   XDSHStack,
   XDSLayout,
