+++ apps/storybook/stories/Divider.stories.tsx
@@ -2,7 +2,8 @@ import type {Meta, StoryObj} from '@storybook/react';
 import * as stylex from '@stylexjs/stylex';
 import {XDSDivider} from '@xds/core/Divider';
 import {XDSCard} from '@xds/core/Card';
-import {XDSVStack, XDSHStack, XDSSection} from '@xds/core/Layout';
+import {XDSSection} from '@xds/core/Section';
+import {XDSVStack, XDSHStack} from '@xds/core/Layout';
 import {XDSText} from '@xds/core/Text';
 import {spacingVars} from '@xds/core/theme/tokens.stylex';
