+++ apps/storybook/stories/Grid.stories.tsx
@@ -2,7 +2,7 @@ import type {Meta, StoryObj} from '@storybook/react';
 import * as stylex from '@stylexjs/stylex';
 import {XDSGrid, XDSGridSpan} from '@xds/core/Grid';
 import {XDSCard} from '@xds/core/Card';
-import {XDSSection} from '@xds/core/Layout';
+import {XDSSection} from '@xds/core/Section';
 import {XDSText} from '@xds/core/Text';
 import {
   colorVars,
