+++ packages/core/src/Center/XDSCenter.tsx
@@ -13,7 +13,7 @@
 import {forwardRef, type HTMLAttributes, type ReactNode} from 'react';
 import * as stylex from '@stylexjs/stylex';
 import type {StyleXStyles} from '@stylexjs/stylex';
-import type {SizeValue} from '../Layout';
+import type {SizeValue} from '../utils/types';
 
 const styles = stylex.create({
   base: {
