+++ packages/core/src/Section/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @file index.ts
+ * @input Imports XDSSection component
+ * @output Exports XDSSection component and types
+ * @position Entry point for @xds/core/Section module
+ *
+ * SYNC: When modified, update /packages/core/src/Section/README.md
+ */
+
+export {XDSSection} from './XDSSection';
+export type {XDSSectionProps, XDSSectionVariant} from './XDSSection';
