+++ packages/core/src/utils/index.ts
@@ -1,9 +1,11 @@
 /**
  * @file index.ts
  * @input Utils module exports
- * @output Re-exports all utility functions
+ * @output Re-exports all utility functions and shared types
  * @position Package entry point for utils
  */
+
+export type {SizeValue} from './types';
 
 export {
   parseDateInput,
