+++ packages/core/src/utils/types.ts
@@ -1,0 +1,11 @@
+/**
+ * @file types.ts
+ * @input None
+ * @output Exports shared utility types
+ * @position Shared types used across multiple components
+ */
+
+/**
+ * Size value type - accepts numbers (treated as pixels) or strings (e.g., '100%', '50vh').
+ */
+export type SizeValue = number | string;
